refactor(scan): split walk git owner
